+++ app/ExternalFiles/Export.php
@@ -372,6 +372,11 @@ class Export extends Tools_Base {
 			return $this->get_not_supported_hint( $listing_obj );
 		}
 
+		// bail if listing object is disabled.
+		if ( $listing_obj->is_disabled() ) {
+			return $this->get_not_enabled_hint( $listing_obj );
+		}
+
 		// get the export object.
 		$export_obj = $listing_obj->get_export_object();
 
@@ -2053,5 +2058,38 @@ class Export extends Tools_Base {
 
 		// return the resulting link.
 		return '<span class="easy-dialog-for-wordpress" data-dialog="' . esc_attr( Helper::get_json( $dialog ) ) . '" title="' . esc_attr__( 'Not supported', 'external-files-in-media-library' ) . '"><span class="dashicons dashicons-editor-help"></span></span>';
+	}
+
+	/**
+	 * Return the "not enabled" hint for table-view.
+	 *
+	 * @param object|false $listing_obj The used service object.
+	 *
+	 * @return string
+	 */
+	private function get_not_enabled_hint( object|false $listing_obj ): string {
+		// bail if object is not an "Extension_Base" object.
+		if ( ! $listing_obj instanceof Service_Base ) {
+			return '';
+		}
+
+		// create the dialog for sync now.
+		$dialog = array(
+			'className' => 'efml',
+			/* translators: %1$s will be replaced by a title. */
+			'title'     => sprintf( __( '%1$s is not enabled', 'external-files-in-media-library' ), $listing_obj->get_label() ),
+			'texts'     => array(
+				/* translators: %1$s will be replaced by a title. */
+				'<p>' . sprintf( __( 'Export to %1$s cannot be used as the service is not available.', 'external-files-in-media-library' ), $listing_obj->get_label() ) . '</p>',
+			),
+			'buttons'   => array(
+				array(
+					'action'  => 'closeDialog();',
+					'variant' => 'primary',
+					'text'    => __( 'OK', 'external-files-in-media-library' ),
+				),
+			),
+		);
+		return '<span class="easy-dialog-for-wordpress" data-dialog="' . esc_attr( Helper::get_json( $dialog ) ) . '" title="' . esc_attr__( 'Not enabled', 'external-files-in-media-library' ) . '"><span class="dashicons dashicons-editor-help"></span></span>';
 	}
 }